fileSize double로 변경 이미지 상세 조회에 날짜 'YYYY-MM-DD'로 변경 이미지 상세 조회 테이블 CSS 변경, fileSize KB 표시

+++ src/main/java/com/roadscanner/service/upload/FileUploadServiceImpl.java
@@ -223,7 +223,7 @@ public class FileUploadServiceImpl implements PcwkLogger, FileUploadService {
 		String datestr = dateFormat.format(currentDate);
 		String fileName = datestr + "_" + file.getOriginalFilename();
 		String url = "https://" + bucketName + ".S3." + region + ".amazonaws.com/" + fileName;
-		int fileSize = (int) (file.getSize() / 1024);
+		double fileSize = file.getSize() / 1024.0;
 
 		File convertedFile = new File(file.getOriginalFilename());
 		try (FileOutputStream fos = new FileOutputStream(convertedFile)) {
